requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pydantic
python-dotenv
asyncio
structlog
google-generativeai
mcp
